Refactors bluetooth handling
Moves bluetooth related code into d39_bt
