+++ CMakeLists.txt
@@ -7,5 +7,9 @@ find_package(Zephyr)
 project(d39_fuel_flow)
 
 # Add application code
-target_sources(app PRIVATE src/main.c src/nmea.c src/d39_gpio/d39_gpio.c)
+target_sources(app PRIVATE 
+                src/main.c
+                src/nmea.c
+                src/d39_gpio/d39_gpio.c
+                src/d39_bt/d39_bt.c)
 target_include_directories(app PRIVATE src)
