+++ src/d39_bt/d39_bt.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "config.h"
+
+int d39_bt_init();
+int d39_bt_send_msg(char *buf, size_t len);
